requirements.txt
dnspython==1.16.0
gunicorn==20.1.0
eventlet==0.31.0
Flask==2.0.1
Flask-Session==0.4.0
Flask-SocketIO==5.1.0
monotonic==1.5
python-dotenv==0.10.1
python-socketio==5.3.0
python-engineio==4.2.0
redis==3.5.3
